requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cached-property==2.0.1
certifi==2025.4.26
charset-normalizer==3.4.2
chevron==0.14.0
decorator==5.2.1
distro==1.9.0
filelock==3.18.0
frozendict==2.4.6
fsspec==2025.5.1
genanki==0.13.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
jiter==0.10.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
moviepy==2.2.1
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai==1.86.0
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
pillow==11.2.1
proglog==0.1.12
pydantic==2.11.6
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
yt-dlp==2025.6.9
